Update omim.ps preferred prefix to OMIMPS in obo context (#1128)
* Update omim.ps preferred prefix to OMIMPS in obo context

Most people in the community have moved to using MIM identifiers for OMIM, which we all should. However, some of (like Mondo) have not managed to do that move yet. This change is safe for anyone using the mim strategy.

* Change order of elements to satisfy linter

Seems the linter does not like the order in my context, so I change it here